+++ changelog.md
@@ -8,6 +8,7 @@
 - Add kwargs to update_entity
 - Call update_entity for ENTITY_EVENTS
 - Ignore click events on plugs
+- Add manufacturer to device
 
 # Version 2023.7.3 (2023-07-13)